refactor: update import statements for consistency and remove unused imports
