+++ packages/vscode/src/state.ts
@@ -8,27 +8,25 @@ import { VSCodeHost } from "./vshost";
 import { applyEdits, toRange } from "./edit";
 import { Utils } from "vscode-uri";
 import { saveAllTextDocuments } from "./fs";
-import { DiagnosticSeverity, parseAnnotations, PromptParameters } from "@genaiscript/core";
-import { Project, PromptScriptRunOptions } from "@genaiscript/core";
-import { ChatCompletionsProgressReport } from "@genaiscript/core";
-import { fixGitHubCopilotInstructions, fixPromptDefinitions } from "@genaiscript/core";
-import { logMeasure } from "@genaiscript/core";
+import { delay, groupBy } from "es-toolkit";
+import { createWebview } from "./webview";
+import { parseAnnotations } from "../../core/src/annotations";
+import { ChatCompletionsProgressReport } from "../../core/src/chattypes";
+import { isEmptyString } from "../../core/src/cleaners";
 import {
   TOOL_NAME,
   CHANGE,
-  TOOL_ID,
   GENAI_ANYTS_REGEX,
+  TOOL_ID,
   MODEL_PROVIDER_GITHUB_COPILOT_CHAT,
-} from "@genaiscript/core";
-import { isCancelError } from "@genaiscript/core";
-import { MarkdownTrace } from "@genaiscript/core";
-import { logInfo, groupBy, logVerbose } from "@genaiscript/core";
-import { GenerationResult } from "@genaiscript/core";
-import { randomHex } from "@genaiscript/core";
-import { delay } from "es-toolkit";
-import { Fragment } from "@genaiscript/core";
-import { createWebview } from "./webview";
-import { isEmptyString } from "@genaiscript/core";
+} from "../../core/src/constants";
+import { randomHex } from "../../core/src/crypto";
+import { isCancelError } from "../../core/src/error";
+import { Fragment } from "../../core/src/generation";
+import { fixPromptDefinitions, fixGitHubCopilotInstructions } from "../../core/src/scripts";
+import { PromptScriptRunOptions, GenerationResult, Project } from "../../core/src/server/messages";
+import { MarkdownTrace } from "../../core/src/trace";
+import { logInfo, logVerbose } from "../../core/src/util";
 
 export const FRAGMENTS_CHANGE = "fragmentsChange";
 export const AI_REQUEST_CHANGE = "aiRequestChange";
