+++ packages/vscode/src/statusbar.ts
@@ -5,7 +5,7 @@ import * as vscode from "vscode";
 import { ExtensionState } from "./state";
 import { toMarkdownString } from "./markdown";
 import { ICON_LOGO_NAME, CHANGE } from "../../core/src/constants";
-import { toStringList } from "../../core/src/utils";
+import { toStringList } from "../../core/src/util";
 
 export function activateStatusBar(state: ExtensionState) {
   const { context } = state;
